Hazmat: fix derp if file is GIF

+++ userbot/modules/hazmat.py
@@ -28,12 +28,20 @@ async def _(hazmat):
     chat = "@hazmat_suit_bot"
     await hazmat.edit("```Suit Up Capt!, We are going to purge some virus...```")
     message_id_to_reply = hazmat.message.reply_to_msg_id
+    msg_reply = None
     async with hazmat.client.conversation(chat) as conv:
         try:
             msg = await conv.send_message(reply_message)
             if level:
                 m = f"/hazmat {level}"
-                msg_level = await conv.send_message(
+                msg_reply = await conv.send_message(
+                          m,
+                          reply_to=msg.id)
+                r = await conv.get_response()
+                response = await conv.get_response()
+            elif reply_message.gif:
+                m = f"/hazmat"
+                msg_reply = await conv.send_message(
                           m,
                           reply_to=msg.id)
                 r = await conv.get_response()
@@ -49,7 +57,7 @@ async def _(hazmat):
             await hazmat.edit("`Can't handle this GIF...`")
             await hazmat.client.delete_messages(
                 conv.chat_id,
-                [msg.id, response.id, r.id, msg_level.id])
+                [msg.id, response.id, r.id, msg_reply.id])
             return
         else:
             downloaded_file_name = await hazmat.client.download_media(
@@ -63,15 +71,13 @@ async def _(hazmat):
                 reply_to=message_id_to_reply
             )
             """ - cleanup chat after completed - """
-            try:
-                msg_level
-            except NameError:
-                await hazmat.client.delete_messages(conv.chat_id,
-                                                 [msg.id, response.id])
-            else:
+            if msg_reply is not None:
                 await hazmat.client.delete_messages(
                     conv.chat_id,
-                    [msg.id, response.id, r.id, msg_level.id])
+                    [msg.id, msg_reply.id, r.id, response.id])
+            else:
+                await hazmat.client.delete_messages(conv.chat_id,
+                                                 [msg.id, response.id])
     await hazmat.delete()
     return os.remove(downloaded_file_name)
